# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gerenciamento de tarefas para disciplina de DEVOPS",
        "title": "Task Manager API",
        "contact": {
            "name": "Robson Gominho",
            "email": "rag2@aluno.ifal.edu.br"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/account/profile": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Rota que retorna todas as informações do usuário logado.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conta do usuário"
                ],
                "summary": "Pesquisar dados do perfil do usuário logado",
                "operationId": "FindProfile",
                "responses": {
                    "200": {
                        "description": "Requisição realizada com sucesso.",
                        "schema": {
                            "$ref": "#/definitions/response.Account"
                        }
                    },
                    "401": {
                        "description": "Usuário não autorizado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Acesso negado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "422": {
                        "description": "Algum dado informado não pôde ser processado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "503": {
                        "description": "A base de dados não está disponível.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Rota que permite que um usuário se autentique no sistema utilizando seu endereço de e-mail e senha.\n| E-mail              | Senha     | Função                                                            |\n|---------------------|-----------|-------------------------------------------------------------------|\n| robson@gmail.com | Test1234! | Usuário do sistema. |",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rotas de autenticação"
                ],
                "summary": "Fazer login no sistema",
                "operationId": "Login",
                "parameters": [
                    {
                        "description": "JSON com todos os dados necessários para que o login seja realizado.",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Requisição realizada com sucesso.",
                        "schema": {
                            "$ref": "#/definitions/response.Authorization"
                        }
                    },
                    "400": {
                        "description": "Requisição mal formulada.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Usuário não autorizado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Acesso negado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "422": {
                        "description": "Algum dado informado não pôde ser processado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "503": {
                        "description": "A base de dados não está disponível.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Rota que permite que um usuário faça logout no sistema.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rotas de autenticação"
                ],
                "summary": "Fazer logout no sistema",
                "operationId": "Logout",
                "responses": {
                    "204": {
                        "description": "Requisição realizada com sucesso."
                    },
                    "400": {
                        "description": "Requisição mal formulada.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Usuário não autorizado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Acesso negado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "422": {
                        "description": "Algum dado informado não pôde ser processado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "503": {
                        "description": "A base de dados não está disponível.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rotas de autenticação"
                ],
                "summary": "Gerar um novo par de tokens para autenticação",
                "operationId": "Refresh",
                "parameters": [
                    {
                        "description": "JSON com todos os dados necessários para que o login seja realizado.",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RefreshDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Requisição realizada com sucesso.",
                        "schema": {
                            "$ref": "#/definitions/response.Authorization"
                        }
                    },
                    "400": {
                        "description": "Requisição mal formulada.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Usuário não autorizado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Acesso negado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "422": {
                        "description": "Algum dado informado não pôde ser processado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "503": {
                        "description": "A base de dados não está disponível.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.LoginDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "robson@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "Test1234!"
                }
            }
        },
        "request.RefreshDTO": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "response.Account": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/response.Role"
                }
            }
        },
        "response.Authorization": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorMessage": {
            "type": "object",
            "properties": {
                "duplicated_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "invalid_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.InvalidField"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "response.InvalidField": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "field_name": {
                    "type": "string"
                }
            }
        },
        "response.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}